requirements.txt
absl-py==2.2.2
accelerate==1.5.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==4.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chromadb==1.0.9
click==8.1.8
cloudpathlib==0.21.1
cmake==3.30.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.10
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.1
ctransformers==0.2.5
cycler==0.12.1
cymem==2.0.11
cytoolz==1.0.1
dataclasses-json==0.5.14
debugpy==1.8.12
decorator==5.1.1
deep-sort-realtime==1.3.2
deepface==0.0.83
Deprecated==1.2.18
diffusers==0.32.2
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.9
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
flask-cors==6.0.0
flatbuffers==25.2.10
floret==0.10.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==4.21.0
gradio_client==0.12.0
graphviz==0.20.3
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.5
jiter==0.8.2
jiwer==3.1.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.10.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain
langchain-community
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langchainplus-sdk
langcodes==3.5.0
langsmith==0.3.1
language_data==1.3.0
lap==0.5.12
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
loguru==0.7.3
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.1.0
murmurhash==1.0.13
mypy-extensions==1.0.0
namex==0.0.9
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai==0.28.0
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
pinecone-client==2.2.4
platformdirs==4.3.6
plotly==6.1.0
pooch==1.8.2
posthog==4.0.1
preshed==3.0.10
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.1
pypdf==5.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
ruff==0.11.10
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.4
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smart-open
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
textacy==0.12.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
typer
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3
ultralytics==8.3.142
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
validators==0.34.0
virtualenv==20.28.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
weasel
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.2
xgboost==3.0.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
